ECC-1822 Refactoring Address Controller
